requirements.txt
colorama
pytest
pylint
ipdb
jupyterlab
numpy
pandas
matplotlib.pyplot
seaborn
scikit-learn
ipykernel
nbresult
pygeohash
yapf
audio-recorder-streamlit
requests
pydub
python-multipart

librosa==0.9.2
#tensorflow==2.10.0

google-cloud-storage==2.4.0
#google-cloud-bigquery<3.0.0

# API
fastapi
pytz
uvicorn
pydantic
